Using koohii cuz yes + adds retardness

+++ core/src/com/dasher/osugdx/IO/Beatmaps/BeatMapStore.java
@@ -6,42 +6,29 @@ import com.badlogic.gdx.Preferences;
 import com.badlogic.gdx.files.FileHandle;
 import com.badlogic.gdx.utils.Array;
 import com.dasher.osugdx.IO.GameIO;
+import com.github.francesco149.koohii.Koohii;
 
 import org.jetbrains.annotations.NotNull;
 
 import java.io.BufferedInputStream;
 import java.io.BufferedOutputStream;
+import java.io.BufferedReader;
 import java.io.EOFException;
 import java.io.File;
 import java.io.FileInputStream;
-import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.io.InputStreamReader;
 import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
-import java.util.Arrays;
-import java.util.Iterator;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.concurrent.CompletableFuture;
-import java.util.stream.Collectors;
-
-import me.zeroeightysix.osureader.OsuFile;
-import me.zeroeightysix.osureader.OsuReader;
-import me.zeroeightysix.osureader.parse.OsuFileParser;
-import me.zeroeightysix.osureader.parse.StandardOsuFileParser;
 
 public class BeatMapStore {
-    public final int VERSION = 1;
+    public final int VERSION = 3;
     private boolean finishedLoadingCache = false;
     private final Array<String> specialFiles = new Array<>();
     private final Array<BeatMapSet> tempCachedBeatmaps = new Array<>();
     private final Array<BeatMapSet> beatMapSets = new Array<>();
-    private boolean beatmapLoadingStarted = false;
     private final FileHandle songsDir;
-    private final Iterator<FileHandle> beatMapSetFoldersIterator;
-    private boolean finishedLoading = false;
-    private final OsuFileParser osuFileParser;
     private boolean libraryChanged = false;
     private final Long beatmapStoreCreationTime;
     public final File libCacheFile;
@@ -54,8 +41,6 @@ public class BeatMapStore {
         String libCacheFilePath = songsDir.path() + "/.beatmap_cache.db";
         beatmapStorePrefs = Gdx.app.getPreferences(getClass().getSimpleName());
         libCacheFile = Gdx.files.external(libCacheFilePath).file();
-        beatMapSetFoldersIterator = new Array<>(songsDir.list()).iterator();
-        osuFileParser = new StandardOsuFileParser();
         beatmapStoreCreationTime = System.nanoTime();
         setupCaching();
         specialFiles.add(libCacheFile.getName());
@@ -75,15 +60,21 @@ public class BeatMapStore {
             verifyVersion();
             try {
                 istream = new ObjectInputStream(new BufferedInputStream(new FileInputStream(libCacheFile)));
+            } catch (EOFException e) {
+                System.out.println("Empty cache!");
+                libraryChanged = true;
+                finishCacheLoading();
             } catch (IOException e) {
                 e.printStackTrace();
                 libraryChanged = true;
                 clearCache();
+                finishCacheLoading();
             }
         } else {
             try {
                 if (libCacheFile.createNewFile()) {
                     System.out.println("Created a new beatmap cache library successfully");
+                    finishCacheLoading();
                 }
             } catch (IOException e) {
                 e.printStackTrace();
@@ -92,11 +83,10 @@ public class BeatMapStore {
     }
 
     private void clearCache() {
+        libraryChanged = true;
         if (libCacheFile.exists()) {
             if (libCacheFile.delete()) {
                 setupCaching();
-                finishCacheLoading();
-                libraryChanged = true;
                 System.out.println("CLEARED LIBRARY CACHE!");
             } else {
                 System.out.println("Failed to clear library cache");
@@ -108,11 +98,23 @@ public class BeatMapStore {
         return !beatMapSet.isEmpty() && beatMapSet.folder.exists() && beatMapSet.folder.isDirectory();
     }
 
-    private void loadCache() {
+    public void loadCache() {
         System.out.println("Started reading cache");
 
         while (!isFinishedLoadingCache()) {
             loadNextCachedBeatmapSet();
+        }
+
+        System.out.println(isFinishedLoadingCache());
+        if (!isFinishedLoadingCache()) {
+            if (istream != null) {
+                try {
+                    istream.close();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            }
+            finishCacheLoading();
         }
     }
 
@@ -120,30 +122,27 @@ public class BeatMapStore {
         if (istream == null) {
             return;
         }
+
+        Object cachedBeatmapSet = null;
         try {
-            Object cachedBeatmapSet = istream.readObject();
-            if (cachedBeatmapSet instanceof BeatMapSet) {
-                BeatMapSet beatMapSet = (BeatMapSet) cachedBeatmapSet;
-                if (isBeatmapSetValid(beatMapSet)) {
-                    beatMapSets.add(beatMapSet);
-                    tempCachedBeatmaps.add(beatMapSet);
-                    for (OsuFile osuFile: beatMapSet) {
-                        osuFile.wasCachedInTheStart = true;
-                        logBeatmapLoaded(osuFile);
-                    }
+            cachedBeatmapSet = istream.readObject();
+        } catch (EOFException e) {
+            finishCacheLoading();
+        } catch (IOException | ClassNotFoundException e) {
+            clearCache();
+            e.printStackTrace();
+        }
+
+        if (cachedBeatmapSet instanceof BeatMapSet) {
+            BeatMapSet beatMapSet = (BeatMapSet) cachedBeatmapSet;
+            if (isBeatmapSetValid(beatMapSet)) {
+                beatMapSets.add(beatMapSet);
+                tempCachedBeatmaps.add(beatMapSet);
+                for (Koohii.Map map : beatMapSet) {
+                    System.out.print("DB CACHE: ");
+                    logBeatmapLoaded(map);
                 }
             }
-        } catch (EOFException e) {
-            System.out.println("Iterated through all maps saved in the db");
-            try {
-                istream.close();
-            } catch (IOException ioException) {
-                ioException.printStackTrace();
-            }
-            finishCacheLoading();
-        } catch (Exception e) {
-            e.printStackTrace();
-            clearCache();
         }
     }
 
@@ -153,10 +152,14 @@ public class BeatMapStore {
     }
 
     private void deleteBeatmapFile(@NotNull File beatmapFile) {
-        if (beatmapFile.delete()) {
-            System.out.println("Deleted the beatmap to save resources");
+        if (beatmapFile.exists()) {
+            if (beatmapFile.delete()) {
+                System.out.println("Deleted the beatmap to save resources");
+            } else {
+                System.out.println("Couldn't deleted beatmap to save resources");
+            }
         } else {
-            System.out.println("Couldn't deleted beatmap to save resources");
+            libraryChanged = true;
         }
     }
 
@@ -170,41 +173,32 @@ public class BeatMapStore {
         }
     }
 
-    private boolean verifyIfMapIsLoadedInCache(BeatMapSet beatMapSet, FileHandle beatmapFile, boolean initialBeatmapLoading) {
+    private boolean verifyIfMapIsLoadedInCache(BeatMapSet beatMapSet, FileHandle beatmapFile) {
         boolean isMapLoadedInCache = false;
         if (beatMapSet == null) {
             return false;
         }
-        if (initialBeatmapLoading) {
-            for (OsuFile file: beatMapSet) {
-                if (file.wasCachedInTheStart) {
+        for (File matchBeatmapFile : beatMapSet.beatmapFiles) {
+            if (matchBeatmapFile.exists()) {
+                if (matchBeatmapFile.equals(beatmapFile.file())) {
                     isMapLoadedInCache = true;
                     break;
                 }
-            }
-        } else {
-            for (File matchBeatmapFile : beatMapSet.beatmapFiles) {
-                if (matchBeatmapFile.exists()) {
-                    if (matchBeatmapFile.equals(beatmapFile.file())) {
-                        isMapLoadedInCache = true;
-                        break;
-                    }
-                } else {
-                    deleteBeatmapFile(matchBeatmapFile);
-                }
+            } else {
+                deleteBeatmapFile(matchBeatmapFile);
             }
         }
         return isMapLoadedInCache;
     }
 
-    private void loadBeatmap(BeatMapSet beatMapSet, BeatMapSet cacheBeatmapSet, @NotNull FileHandle beatmapFile, boolean initialBeatmapLoading) {
+    private void loadBeatmap(BeatMapSet beatMapSet, BeatMapSet cacheBeatmapSet, @NotNull FileHandle beatmapFile) {
         if (!beatmapFile.extension().equals("osu")) {
             return;
         }
 
         File javaBeatmapFile = beatmapFile.file();
-        boolean isMapLoadedInCache = verifyIfMapIsLoadedInCache(cacheBeatmapSet, beatmapFile, initialBeatmapLoading);
-        OsuFile beatMap = null;
+        boolean isMapLoadedInCache = verifyIfMapIsLoadedInCache(cacheBeatmapSet, beatmapFile);
+        Koohii.Map beatMap = null;
 
         if (isMapLoadedInCache) {
             return;
@@ -220,11 +214,22 @@ public class BeatMapStore {
         }
 
         try {
-            beatMap = OsuReader.read(javaBeatmapFile, osuFileParser);
-        } catch (Exception e) {
+            BufferedReader beatmapBR = null;
+            try {
+                beatmapBR = new BufferedReader(new InputStreamReader(new FileInputStream(javaBeatmapFile)));
+                beatMap = new Koohii.Parser().map(beatmapBR);
+            } catch (UnsupportedOperationException e) {
+                System.out.println("Can't parse other beatmap modes...");
+                beatmapBR.close();
+                deleteBeatmapFile(javaBeatmapFile);
+            } catch (Exception e) {
+                e.printStackTrace();
+                System.out.println("Error loading beatmap: " + javaBeatmapFile.getPath());
+                beatmapBR.close();
+                deleteBeatmapFile(javaBeatmapFile);
+            }
+        } catch (IOException e) {
             e.printStackTrace();
-            System.out.println("Error loading beatmap: " + javaBeatmapFile.getPath());
-            deleteBeatmapFile(javaBeatmapFile);
         }
 
         if (beatMap != null) {
@@ -234,12 +239,9 @@ public class BeatMapStore {
         }
     }
 
-    private void logBeatmapLoaded(OsuFile beatMap) {
+    private void logBeatmapLoaded(Koohii.Map beatMap) {
         try {
-            System.out.println(
-                    "Loaded new beatmap: "
-                            + beatMap.getMetadata().getTitle()
-                            + " [" + beatMap.getMetadata().getVersion() + "]");
+            System.out.println("Loaded new beatmap: " + beatMap.title + " [" + beatMap.version + "]");
         } catch (Exception ignore) {}
     }
 
@@ -268,7 +270,7 @@ public class BeatMapStore {
         BeatMapSet beatMapSet = samePathBeatmapSet == null? new BeatMapSet(beatMapSetFolder) : samePathBeatmapSet;
 
         for (FileHandle file : list) {
-           loadBeatmap(beatMapSet, samePathBeatmapSet, file, true);
+           loadBeatmap(beatMapSet, samePathBeatmapSet, file);
         }
 
         addNewBeatmapSet(beatMapSet, samePathBeatmapSet, beatMapSetFolder);
@@ -291,12 +293,13 @@ public class BeatMapStore {
     }
 
     private void loadFirstSteps() {
-        beatmapLoadingStarted = true;
+        System.out.println("Started loading beatmaps...");
     }
 
     private void onLibraryChange() {
         System.out.println("Changes to beatmap library detected!");
         beatmapStorePrefs.putInteger("VERSION", VERSION);
+        beatmapStorePrefs.flush();
         for (BeatMapSet beatMapSet: tempCachedBeatmaps) {
             saveBeatmapSetInCache(beatMapSet);
         }
@@ -315,7 +318,6 @@ public class BeatMapStore {
                 ((System.nanoTime() - beatmapStoreCreationTime) / 1e6) + "ms to load " +
                         beatMapSets.size + " beatmapSets"
         );
-        finishedLoading = true;
         if (ostream != null) {
             try {
                 ostream.close();
@@ -323,38 +325,17 @@ public class BeatMapStore {
                 e.printStackTrace();
             }
         }
-
     }
 
     public void loadAllBeatmaps() {
         loadFirstSteps();
-        for (FileHandle fileHandle : songsDir.list()) {
-            loadBeatmapSet(fileHandle);
+        for (FileHandle file : songsDir.list()) {
+            loadBeatmapSet(file);
         }
         loadFinalSteps();
     }
 
-    public void loadNextBeatMapSet() {
-        loadFirstSteps();
-        if (finishedLoading) {
-            return;
-        }
-        if (beatMapSetFoldersIterator.hasNext()) {
-            loadBeatmapSet(beatMapSetFoldersIterator.next());
-        } else {
-           loadFinalSteps();
-        }
-    }
-
-    public boolean isFinishedLoading() {
-        return finishedLoading;
-    }
-
     public boolean isFinishedLoadingCache() {
         return finishedLoadingCache;
-    }
-
-    public boolean isBeatmapLoadingStarted() {
-        return beatmapLoadingStarted;
     }
 }
